modified nov 2023 task02 to work with chuck's server

+++ Nov 2023/task02/src/shopping/Compare.java
@@ -8,9 +8,10 @@ public class Compare implements Comparator<Product>{
     public int compare(Product p1, Product p2) {
         //System.out.println("in compare");
         if (p1.getRating() != p2.getRating()) {
-            return p2.getRating() - p1.getRating();
+            
+            return (int)(p2.getRating() - p1.getRating());
         } else {
-            return p2.getPrice() - p1.getPrice();
+            return (int)(p2.getPrice() - p1.getPrice());
         }
     }
